requirements.txt
# requirements.txt
PySide6
transitions
pykka
rxpy
pytest
mock
flake8
sphinx
pyinstaller
pyserial
